requirements.txt
pandas>=1.5
numpy>=1.23
scikit-learn>=1.3
pytorch-lightning>=2.1
transformers>=4.36
torch>=2.1
sentencepiece>=0.1.99
nltk>=3.8
spacy>=3.7
textblob>=0.17.1
vaderSentiment>=3.3.2
newsapi-python>=0.2.7
yfinance>=0.2.37
alpha_vantage>=2.3.1
tweepy>=4.14
praw>=7.7
sqlalchemy>=2.0
psycopg2-binary>=2.9
pydantic>=2.6
pydantic-settings>=2.1
python-dotenv>=1.0
httpx>=0.27
beautifulsoup4>=4.12
lxml>=5.1
streamlit>=1.31
plotly>=5.19
matplotlib>=3.8
backoff>=2.2
redis>=5.0
loguru>=0.7
pyyaml>=6.0
